requirements.txt
selenium
ordered_set
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
distro==1.8.0
exceptiongroup==1.1.3
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
instructor
openai==1.1.2
Pillow==10.1.0
playwright==1.39.0
pydantic==2.4.2
pydantic_core==2.10.1
pyee==11.0.1
python-dotenv==1.0.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0